storage settings component refactor
