+++ src/renderer/src/components/settings/storagesettings/index.ts
@@ -1,0 +1,3 @@
+export { StorageStatistics } from './StorageStatistics';
+export { DataManagement } from './DataManagement';
+export { useStorageSettings } from './useStorageSettings';
